backend/requirements.txt
# Core Framework
fastapi==0.115.12
uvicorn==0.30.1
python-multipart==0.0.22
httpx==0.27.0
starlette==0.46.2

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.48
alembic==1.13.1

# Redis
redis==5.0.6

# RabbitMQ
aio-pika==9.4.1

# Auth
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
cryptography==46.0.5

# Validation
pydantic==2.12.5
pydantic-settings==2.3.1
email-validator==2.3.0

# Observability — Prometheus
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.24.1

# Observability — OpenTelemetry
opentelemetry-api==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-instrumentation-fastapi==0.61b0
opentelemetry-instrumentation-asyncpg==0.61b0
opentelemetry-instrumentation-redis==0.61b0

# Logging
structlog==24.2.0

# Async
anyio==4.12.1
aiohttp==3.13.4

# SSE
sse-starlette==2.2.1

# Performance
uvloop==0.21.0
orjson==3.11.7
ujson==5.11.0

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0

# Utilities
python-dotenv==1.2.2
PyYAML==6.0.3